requirements.txt
flask
flask-sqlalchemy
flask-socketio
eventlet
pytest
pytest-flask
